+++ statistics.c
@@ -28,6 +28,21 @@ print_ring_statistics (struct ring *ring)
 	   conflicts, average (conflicts, search));
   PRINTLN ("%-22s %17" PRIu64 " %13.2f per second", "decisions:",
 	   decisions, average (decisions, search));
+  PRINTLN ("  %-22s %17" PRIu64 " %13.2f %% decisions",
+           "heap-decisions:", s->decisions.heap,
+	   percent (s->decisions.heap, decisions));
+  PRINTLN ("  %-22s %17" PRIu64 " %13.2f %% decisions",
+           "negative-decisions:", s->decisions.negative,
+	   percent (s->decisions.negative, decisions));
+  PRINTLN ("  %-22s %17" PRIu64 " %13.2f %% decisions",
+           "positive-decisions:", s->decisions.positive,
+	   percent (s->decisions.positive, decisions));
+  PRINTLN ("  %-22s %17" PRIu64 " %13.2f %% decisions",
+           "queue-decisions:", s->decisions.queue,
+	   percent (s->decisions.queue, decisions));
+  PRINTLN ("  %-22s %17" PRIu64 " %13.2f %% decisions",
+           "random-decisions:", s->decisions.random,
+	   percent (s->decisions.random, decisions));
   PRINTLN ("%-22s %17u %13.2f %% variables", "solving-fixed:",
 	   s->fixed, percent (s->fixed, variables));
   PRINTLN ("%-22s %17u %13.2f %% variables", "failed-literals:",